+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.5.0](https://github.com/alpamayo-solutions/colca/compare/v0.4.0...v0.5.0) (2026-09-19)
+
+
+### ⚠ BREAKING CHANGES
+
+* **alarms:** a publisher that silenced an alarm through an alarm_acknowledgement edit must send a _CmdOperate to the alarm's evaluator.
+
+### Features
+
+* **alarms:** an operator's act is a command to the evaluator, not an edit ([#31](https://github.com/alpamayo-solutions/colca/issues/31)) ([c317950](https://github.com/alpamayo-solutions/colca/commit/c317950f55ae349f42942f201550cd373f1186bb))
+
+## [0.4.0](https://github.com/alpamayo-solutions/colca/compare/v0.3.0...v0.4.0) (2026-09-19)
+
+
+### Features
+
+* **door:** Tombstone retires a record at a topic ([#29](https://github.com/alpamayo-solutions/colca/issues/29)) ([3e1fc72](https://github.com/alpamayo-solutions/colca/commit/3e1fc72a144c86b1af6ee8f04a18c83364e91322))
+
 ## [0.3.0](https://github.com/alpamayo-solutions/colca/compare/v0.2.1...v0.3.0) (2026-09-18)